requirements.txt
pyserial==3.0.1
paho-mqtt==1.1
